Modify demo package.
