+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.chenenyu.router.demo"
+<manifest package="com.chenenyu.router.app"
           xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
