+++ app/src/main/java/com/chenenyu/router/app/App.java
@@ -1,4 +1,4 @@
-package com.chenenyu.router.demo;
+package com.chenenyu.router.app;
 
 import android.support.multidex.MultiDexApplication;
