+++ app/src/main/java/com/chenenyu/router/app/ImplicitActivity.java
@@ -1,18 +1,14 @@
-package com.chenenyu.router.demo;
+package com.chenenyu.router.app;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.widget.TextView;
 
-import com.chenenyu.router.annotation.Route;
-
-@Route({"test", "http://example.com/user"})
-public class TestActivity extends AppCompatActivity {
-
+public class ImplicitActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_test);
+        setContentView(R.layout.activity_scheme);
 
         TextView text = (TextView) findViewById(R.id.text_test);
         Bundle bundle = getIntent().getExtras();
@@ -25,5 +21,6 @@ public class TestActivity extends AppCompatActivity {
                     .append(bundle.getString("status"));
             text.setText(sb.toString());
         }
+
     }
 }
