+++ app/src/main/java/com/chenenyu/router/app/InterceptedActivity.java
@@ -1,4 +1,4 @@
-package com.chenenyu.router.demo;
+package com.chenenyu.router.app;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
