+++ app/src/main/java/com/chenenyu/router/app/TestActivity.java
@@ -1,14 +1,18 @@
-package com.chenenyu.router.demo;
+package com.chenenyu.router.app;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.widget.TextView;
 
-public class ImplicitActivity extends AppCompatActivity {
+import com.chenenyu.router.annotation.Route;
+
+@Route({"test", "http://example.com/user"})
+public class TestActivity extends AppCompatActivity {
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_scheme);
+        setContentView(R.layout.activity_test);
 
         TextView text = (TextView) findViewById(R.id.text_test);
         Bundle bundle = getIntent().getExtras();
@@ -21,6 +25,5 @@ public class ImplicitActivity extends AppCompatActivity {
                     .append(bundle.getString("status"));
             text.setText(sb.toString());
         }
-
     }
 }
