+++ app/src/main/res/layout/activity_intercepted.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.chenenyu.router.demo.InterceptedActivity">
+    tools:context="com.chenenyu.router.app.InterceptedActivity">
 
 </FrameLayout>
