+++ app/src/test/java/com/chenenyu/router/app/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.chenenyu.router.demo;
+package com.chenenyu.router.app;
 
 import org.junit.Test;
